+++ src/Facades/Viva.php
@@ -9,6 +9,7 @@ use Illuminate\Support\Facades\Facade;
  * @method static \Sebdesign\VivaPayments\Services\Order orders()
  * @method static \Sebdesign\VivaPayments\Services\Transaction transactions()
  * @method static \Sebdesign\VivaPayments\Services\Webhook webhooks()
+ * @method static \Sebdesign\VivaPayments\Services\ISV isv()
  * @method static \Sebdesign\VivaPayments\Client withEnvironment(\Sebdesign\VivaPayments\Enums\Environment|string $environment)
  * @method static \Sebdesign\VivaPayments\Client withBasicAuthCredentials(string $merchantId, string $apiKey)
  * @method static \Sebdesign\VivaPayments\Client withOAuthCredentials(string $clientId, string $clientSecret)