+++ src/Services/ISV.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Sebdesign\VivaPayments\Services;
+
+use Sebdesign\VivaPayments\Client;
+
+class ISV
+{
+    public function __construct(protected Client $client)
+    {
+    }
+
+    public function orders(): ISV\Order
+    {
+        return new ISV\Order($this->client);
+    }
+}